requirements.txt
aiogram
openai
python-dotenv
httpx